[REFACTOR][PYTHON] Convert tvm.support module to package; extract libinfo; phase out FrontendTestModule
tvm.support is growing to host load-bearing compiler/CLI/process helpers
moved from tvm.contrib. Converting it from a single-file module to a package
is a prerequisite. libinfo() is moved to its own support/libinfo.py for
single-responsibility; FrontendTestModule is dropped (audit confirmed it has
zero callers outside its own definition — the class was never used by
production code or tests).

The package docstring establishes the inclusion rule distinguishing
tvm.support (load-bearing CLI/host-tool helpers) from tvm.contrib
(optional vendor SDK integrations and experimental features).
